+++ entities/Item2D.java
@@ -5,7 +5,7 @@ public class Item2D extends Item {
 
     private int width;
 
-    public Item2D(String id, double weight, int length, int width) {
+    public Item2D(String id, float weight, int length, int width) {
         super(id, weight, length);
         this.width = width;
     }
